Excepciones-errors
Creando clases de error para utilizarlas como excepciones en las diferentes instanciaciones de los modulos

+++ modulos/club/controllers/clubController.js
@@ -33,6 +33,9 @@ module.exports = class ClubController extends AbstractController {
 
   async clubDetails(req, res) {
     const equipoId = req.params.id;
+    if (equipoId === undefined) {
+      throw new ClubIdNotDefinedError();
+    }
     const equipo = await this.clubService.getById(equipoId);
     try {
       res.render('crudClubes/ver', {
@@ -80,6 +83,9 @@ module.exports = class ClubController extends AbstractController {
 
   clubUpdateGet(req, res) {
     const equipoId = req.params.id;
+    if (equipoId === undefined) {
+      throw new ClubIdNotDefinedError();
+    }
     try {
       res.render('crudClubes/editar', {
         layout: 'index',
@@ -93,6 +99,9 @@ module.exports = class ClubController extends AbstractController {
 
   async clubUpdatePut(req, res) {
     const equipoId = req.params.id;
+    if (equipoId === undefined) {
+      throw new ClubIdNotDefinedError();
+    }
     try {
       const { name, area_name, address } = req.body;
       if (!name || !area_name || !address) {
@@ -116,9 +125,9 @@ module.exports = class ClubController extends AbstractController {
 
   async clubDelete(req, res) {
     try {
-      await this.clubService.deleteEquipo(equipoId);
-      req.session.messages = [`El equipo con id ${equipoId} se elimino correctamente!`];
-      res.redirect('/crudClubes');
+      const { id } = req.params;
+      if (id === undefined) {
+        throw new ClubIdNotDefinedError();
       }
       await this.clubService.deleteEquipo(id);
       req.session.messages = [`El equipo con id ${id} se elimino correctamente!`];
